@@ -3,15 +3,15 @@ namespace SeaDrip\Abstracts;
 
 abstract class FormatRule
 {
-    abstract protected function get_columns() : array;
+    abstract protected function get_columns(): array;
 
-    public static function create() : FormatRule {
-        $clsName = static::class;
-        return new $clsName;
+    public static function create(): static
+    {
+        return new (static::class);
     }
 
-    public function get_alias_map() : array {
+    public function get_alias_map(): array
+    {
         return [];
     }
 }
-
